remove svg.path dependency rename module svg to svg_load path2polygonPoints expects a Path object now instead of a string load_svg now returns a Path object instead of a string
