+++ pysrc/requirements.txt
@@ -1,3 +1,2 @@
-svg.path==3.0
 svgpathtools==1.3.3
 numpy
